Changed setThermalizaztionsteps() variable type from int to unsigned long long int, added more information for effective_mass and gs_energy estimators when results are printed to console and file, improved readability of Metropolis algorithm transition probability

+++ Diagram.hpp
@@ -42,7 +42,7 @@ class Diagram {
         };
 
         // setters
-        void setRelaxSteps(int relax_steps);
+        void setRelaxSteps(unsigned long long int relax_steps);
 
         // getters
         inline long long unsigned int getNdiags() const {return _N_diags;};
